extract ARSCData and FlagsOffset from ARSCDecoder and remove unused imports
