+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/ApkBuilder.java
@@ -19,7 +19,6 @@ package brut.androlib;
 import brut.androlib.exceptions.AndrolibException;
 import brut.androlib.apk.ApkInfo;
 import brut.androlib.apk.UsesFramework;
-import brut.androlib.res.ResourcesDecoder;
 import brut.androlib.res.Framework;
 import brut.androlib.res.data.ResConfigFlags;
 import brut.androlib.res.xml.ResXmlPatcher;
