+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/res/Framework.java
@@ -19,6 +19,8 @@ package brut.androlib.res;
 import brut.androlib.Config;
 import brut.androlib.exceptions.AndrolibException;
 import brut.androlib.exceptions.CantFindFrameworkResException;
+import brut.androlib.res.data.ARSCData;
+import brut.androlib.res.data.FlagsOffset;
 import brut.androlib.res.decoder.ARSCDecoder;
 import brut.util.Jar;
 import org.apache.commons.io.IOUtils;
@@ -63,7 +65,7 @@ public class Framework {
             in = zip.getInputStream(entry);
             byte[] data = IOUtils.toByteArray(in);
 
-            ARSCDecoder.ARSCData arsc = ARSCDecoder.decode(new ByteArrayInputStream(data), true, true);
+            ARSCData arsc = ARSCDecoder.decode(new ByteArrayInputStream(data), true, true);
             publicizeResources(data, arsc.getFlagsOffsets());
 
             File outFile = new File(getFrameworkDirectory(), arsc
@@ -140,8 +142,8 @@ public class Framework {
         publicizeResources(arsc, ARSCDecoder.decode(new ByteArrayInputStream(arsc), true, true).getFlagsOffsets());
     }
 
-    private void publicizeResources(byte[] arsc, ARSCDecoder.FlagsOffset[] flagsOffsets) {
-        for (ARSCDecoder.FlagsOffset flags : flagsOffsets) {
+    private void publicizeResources(byte[] arsc, FlagsOffset[] flagsOffsets) {
+        for (FlagsOffset flags : flagsOffsets) {
             int offset = flags.offset + 3;
             int end = offset + 4 * flags.count;
             while (offset < end) {
